fix achievement sounds in the macsound interface
> With sounds on, using #levelchange to drop more than one level only
> gave the level-loss sound effect once.

Fix the underlying issue within macsound.m by allowing the currently
playing sound to run to its completion prior to starting the next.

This also backs out the acknowledgment --More-- that was added after
the sound issue first materialized and puts it back to the way it was
the past several releases.