fixes entry for pull request #993 - curses' >>
Pull request from entrez:  when the curses interface was deciding
whether ">>" could fit on the bottom line of the message window, it
was off by two (based on the initial report) and ">>" clobbered the
last two characters.

Reported case was "Welcome to level " when that was appended to
another short message.  It should have ended with "6. " but that had
been chopped off by ">> ".  Reproducible by assigning a long name to
something, dropping an item, dropping the something, and dropping
just enough other stuff so that the named item will be reported on
the bottom message line during pickup-all, forcing --More-- before
continuing to the item below it.  The item with the long name needs
to fit in the message window's width (when formatted with invent
letter prefix and trailing period) without being wrapped and not
leaving enough unused space to fix ">> ".

Closes #993