parent 3cb2eb4a026e7a83fab3ae052558f2e4ad2a17b1 author Ray Chason <ray.chason@protonmail.com> 1684372172 -0400 committer nhmall <nhmall@nethack.org> 1685414340 -0400
Add configuration to support Curses on WinGUI

and enable support for Unicode on Curses.

Allow NetHackW to select the Curses interface

Reorder drawing of extended command prompt

Curses on WinGUI needs this change. This may be a bug in PDCursesMod,
but it seems to be harmless to the other ports.

Avoid calling Curses after the windows are closed

Provide erase_char and kill_char for WinGUI Curses

Set Lua version to 5.4.6