give fuzzer some protection against brainlessness
With the LifeSaved property (via amulet), hero being killed by
brainlessness gets killed twice.  But for explore|wizard mode where
the answer to "Die?" might be "no" and for the fuzzer where it's
always "no", a mind flayer's 3 drain-Int attacks or master flayer's
5 drain-Int attacks will usually kill the hero all over again (and
again, ...).  Skip remaining ones, like happens when one of them
hits and discovers that the target has no head or is mindless, for
the rest of the flayer's current move.

Monsters wearing life-saving don't get killed twice.  That doesn't
seem very fair, but this hasn't touched that.